__dev/fluentd/requirements.txt
# coding:utf-8
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

fluent-logger==0.9.3
pymongo==3.6.1